Refactor : 2차 인증 폴더 이름 변경
